Print anomaly type with DLI_IDENTIFY_2 on Deneva
